refactor: Applying the SRP in the Home page
