+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { ConfigProvider, Layout } from 'antd'
-import Home from './pages/Home'
+import Home from './pages/home/Home'
 
 function App() {
   return (
